requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.7.0
albumentations>=1.1.0
timm>=0.6.0
segmentation-models-pytorch>=0.3.0
torchmetrics>=0.6.0

# For TensorRT optimization
tensorrt>=8.0.0
pycuda>=2021.1
onnx>=1.10.0
onnxruntime-gpu>=1.9.0

# For model compression
torch-pruning>=1.0.0
# pytorch-quantization>=2.1.0

# Utilities
pyyaml>=5.4.0
hydra-core>=1.1.0
wandb>=0.12.0
rich>=10.0.0
psutil>=5.8.0